refactoring according to the last comments
